requirements.txt
django==3.1.3
djangorestframework==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
requests==2.24.0
virtualenv==20.1.0